+++ R/plot-intensity-map.R
@@ -29,23 +29,29 @@ plot_intensity_map <- function(
     )
     extra_layers <- remove_null_elements(extra_layers)
 
-    if (options$intensity_map$density) {
-        ggplot(dataset, aes(x = .data$rt, y = .data$mz)) +
+    x_dim <- options$intensity_maps$x_dim
+    y_dim <- options$intensity_maps$y_dim
+    fill_scale <- options$intensity_maps$fill_scale
+
+    dim_label <- function(dim) if (dim == "rt") "RT (sec)" else "m/z"
+
+    if (options$intensity_maps$density) {
+        ggplot(dataset, aes(x = .data[[x_dim]], y = .data[[y_dim]])) +
             geom_density_2d_filled(contour_var = "ndensity") +
-            scale_fill_viridis_d() +
-            labs(x = "RT (sec)", y = "m/z", fill = "Density") +
+            (if (!is.null(fill_scale)) fill_scale else scale_fill_viridis_d()) +
+            labs(x = dim_label(x_dim), y = dim_label(y_dim), fill = "Density") +
             theme_minimal() +
             extra_layers
     } else {
         ggplot(
             dataset,
             aes(
-                x = .data$rt,
-                y = .data$mz,
+                x = .data[[x_dim]],
+                y = .data[[y_dim]],
                 fill = log1p(.data$intensity))) +
             geom_tile() +
-            scale_fill_viridis_c() +
-            labs(x = "RT (sec)", y = "m/z", fill = "log(Intensity)") +
+            (if (!is.null(fill_scale)) fill_scale else scale_fill_viridis_c()) +
+            labs(x = dim_label(x_dim), y = dim_label(y_dim), fill = "log(Intensity)") +
             theme_minimal() +
             extra_layers
     }